TweetSentimentProcessing/requirements.txt
bert-for-tf2
sentencepiece
numpy
pandas
tensorflow
tensorflow-hub
beautifulsoup4
tqdm